Revert "5 controllers"

+++ src/main/java/com/tgt/rysetii/learningresourcesapi/service/LearningResourceService.java
@@ -31,25 +31,36 @@ public class LearningResourceService {
     public List<Double> getProfitMargin()
     {
         List<LearningResource> learningResources=getLearningResources();
-        List<Double> profitMargins=learningResources.stream()
-                .map(lr->((lr.getSellingPrice()-lr.getCostPrice())/lr.getSellingPrice())).collect(toList());
+
+            List<Double> profitMargins;
+        profitMargins = new ArrayList<>();
+        for(LearningResource lr:learningResources)
+            {
+                double r=(lr.getSellingPrice() - lr.getCostPrice())/lr.getSellingPrice();
+                profitMargins.add(r);
+            }
+
+
         return profitMargins;
 
     }
     public List<LearningResource> sortLearningResources()
     {
-        List<LearningResource> learningResources=getLearningResources();
-        learningResources.sort((lr1,lr2)->{
-            Double profitMargin1=(lr1.getSellingPrice()-lr1.getCostPrice())/lr1.getSellingPrice();
-            Double profitMargin2=(lr1.getSellingPrice()-lr1.getCostPrice())/lr1.getSellingPrice();
-            return profitMargin2.compareTo(profitMargin1);
-        });
-        return learningResources;
+        List<LearningResource> lr=getLearningResources();
+        Collections.sort(lr,new LearningResourceSorter());
+        return lr;
     }
-    public void deleteLearningResource(int id)
+    public String deleteLearningResource(int id)
     {
-        
+        if(a.existsById(id))
+        {
             a.deleteById(id);
+            return "Deleted the Learning Resource Successfully";
+        }
+        else
+        {
+            return "Learning Resource Doesn't Exist!!";
+        }
     }
 
 }
